added home page to mobile
